createTreeView relocated to antd to core
