--- packages/antd/src/definitions/treeView/index.ts
@@ -1,1 +1,0 @@
-export { createTreeView } from "./createTreeView/createTreeView";
