+++ packages/antd/src/index.tsx
@@ -49,4 +49,3 @@ export {
     mapAntdFilterToCrudFilter,
 } from "./definitions/table";
 export { getValueFromEvent } from "./definitions/upload";
-export { createTreeView } from "./definitions/treeView";
