+++ packages/core/src/definitions/helpers/index.ts
@@ -4,3 +4,4 @@ export { userFriendlyResourceName } from "./userFriendlyResourceName";
 export { handleUseParams } from "./handleUseParams";
 export { queryKeys } from "./queryKeys";
 export { routeGenerator } from "./routeGenerator";
+export { createTreeView } from "./treeView/createTreeView";
