+++ packages/core/src/index.tsx
@@ -67,5 +67,6 @@ export {
     importCSVMapper,
     handleUseParams,
     routeGenerator,
+    createTreeView,
 } from "./definitions/helpers";
 export { file2Base64 } from "./definitions/upload";
