+++ packages/core/src/interfaces/index.ts
@@ -46,6 +46,9 @@ export * from "./metaData";
 //queryKeys
 export * from "./queryKey";
 
+//menuItem
+export * from "./menuItem";
+
 //metaData
 export * from "./live";
