+++ packages/core/src/interfaces/menuItem.ts
@@ -1,0 +1,10 @@
+import { IResourceItem } from "src";
+
+export type ITreeMenu = IMenuItem & {
+    children: ITreeMenu[];
+};
+
+export type IMenuItem = IResourceItem & {
+    key: string;
+    route: string;
+};
